% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLR.R
\name{expit}
\alias{expit}
\title{expit a log odds to obtain a probability}
\usage{
expit(x)
}
\arguments{
\item{x}{a log odds}
}
\value{
a probability
}
\description{
expit a log odds to obtain a probability.  This is the antilogit function
}
