% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLR.R
\name{SOLR}
\alias{SOLR}
\title{Sequential Offsetted (logistic) Regression}
\usage{
SOLR(ymod, zmod, id, pi1.pi0.ratio, DAT.ods, CORSTR = "independence")
}
\arguments{
\item{ymod}{The target response model}

\item{zmod}{The auxiliary variable model}

\item{id}{Subject identifier}

\item{pi1.pi0.ratio}{The ratio of sampling probabilities pi(z=1, X)/pi(Z=0, X)}

\item{DAT.ods}{The dataframe}

\item{CORSTR}{The correlation structure.  Presently only allows "independence", "exchangeabel" and "ar1".
Note that one should only used "independence" if using observation level sampling.}
}
\value{
beta: Parameter estimates for the target model.

robvarbeta: Robust covariance matrix for the target model

gamma: Parameter estimates for the auxiliary variable model

robvargamma: Robust covariance matrix for the auxiliary variable model
}
\description{
Conducts sequential offsetted logistic regression analyses
}
